modules/platform/repos/cosmos-poc/requirements.txt
openai
anthropic
httpx
requests
pytest
pytest-asyncio
python-dotenv
Pillow
numpy
opencv-python
python-socketio[asyncio_client]
